Added custom iterators for instructions, basic blocks, and uses of instructions. Also split Instruction and OpCode into separate files
